:broom: lint: from eslint and prettier to biome
