--- tests/E2E/user-register.test.ts
@@ -1,25 +1,0 @@
-import request from 'supertest'
-import { afterAll, beforeAll, describe, expect, it } from 'vitest'
-
-import { app } from '@/app'
-
-describe('User register (e2e)', () => {
-  beforeAll(async () => {
-    await app.ready()
-  })
-
-  afterAll(async () => {
-    await app.close()
-  })
-
-  it('should be able to register', async () => {
-    const response = await request(app.server).post('/users').send({
-      name: 'John Doe',
-      serviceGender: 'Male',
-      email: 'john.doe@gmail.com',
-      password: '123456',
-    })
-
-    expect(response.statusCode).toEqual(201)
-  })
-})
