+++ tests/e2e/user-profile.spec.ts
@@ -1,0 +1,31 @@
+import request from 'supertest'
+import { afterAll, beforeAll, describe, expect, it } from 'vitest'
+
+import { app } from '@/app'
+import { createAndAuthenticateUser } from '@/utils/create-and-authenticate-user'
+
+describe('User profile (e2e)', () => {
+	beforeAll(async () => {
+		await app.ready()
+	})
+
+	afterAll(async () => {
+		await app.close()
+	})
+
+	it('should be able to get user profile', async () => {
+		const { token } = await createAndAuthenticateUser(app)
+
+		const profileResponse = await request(app.server)
+			.get('/me')
+			.set('Authorization', `Bearer ${token}`)
+			.send()
+
+		expect(profileResponse.statusCode).toEqual(200)
+		expect(profileResponse.body.user).toEqual(
+			expect.objectContaining({
+				email: 'john.doe@gmail.com',
+			}),
+		)
+	})
+})
